% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{allsimp.emmc}
\alias{allsimp.emmc}
\title{Custom Contrast Function for emmeans}
\usage{
allsimp.emmc(levs, ...)
}
\arguments{
\item{levs}{Interaction levels (should be of length 8).}
}
\description{
Create a contrast matrix for equivalence test of a 2x4
interaction.
}
\details{
Runs all six ways of comparing the simple effects of
the two-level factor. For use with the emmeans package.
}
