% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{check_fake}
\alias{check_fake}
\title{Is Data Flagged As Simulated?}
\usage{
check_fake(path)
}
\arguments{
\item{path}{Name of the subdirectory.}
}
\description{
Check whether the data in the subdirectory is flagged as simulated.
}
