% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cjudgments}
\alias{cjudgments}
\title{Category Judgments from Truth Trajectory Study}
\format{
A data frame (tibble) with 39,406 observations on 3 variables:
\describe{
  \item{ID}{Participant identifier.}
  \item{stim_id}{Stimulus identifier.}
  \item{category}{Category.}
}
}
\usage{
cjudgments
}
\description{
A dataset representing category judgments for each statement by
each participant during the exposure phase (phase 1).
}
\keyword{datasets}
