% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clmm_maximal}
\alias{clmm_maximal}
\title{Fitted Cumulative Link Mixed Model for Nadarevic and Erdfelder Data}
\format{
An object of class "clmm", resulting from a call to the
\code{\link[ordinal]{clmm}} function.
}
\usage{
clmm_maximal
}
\description{
Fitted Cumulative Link Mixed Model for Nadarevic and Erdfelder Data
}
\details{
The object is the result of the function call

\code{ordinal::clmm(trating ~ R * D + (R * D | subj_id) +
                                      (R * D | item_id), NE_exp1)}.

The fitted model is stored as an independent object in the package
because the fitting process is too slow to allow it to be
re-created whenever it is needed.
}
\seealso{
\code{\link{NE_exp1}}
}
\keyword{datasets}
