% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{codebook}
\alias{codebook}
\title{Compile and Display Codebook and Materials}
\usage{
codebook(show_stim = TRUE, browse = TRUE)
}
\arguments{
\item{show_stim}{Whether to include the stimulus materials.}

\item{browse}{Whether to open the codebook in a browser. Otherwise, it just prints the filename.}
}
\value{
Path to the codebook file.
}
\description{
Compile and display the codebook for anonymized data and stimulus
materials.
}
