% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_gamm}
\alias{fit_gamm}
\title{Fit generalized additive mixed model to ratings data}
\usage{
fit_gamm(.data, main_effect = FALSE)
}
\arguments{
\item{.data}{Data frame, with the format as resulting from a call
to \code{\link{gen_data}}.}

\item{main_effect}{Whether to test the main effect of repetition
(TRUE) or the repetition-by-interval interaction (FALSE; the
default).}
}
\value{
A vector, with the following elements.

\describe{
  \item{\code{R}}{Fixed-effects estimate of the main effect of repetition.}
  \item{\code{I1}}{Fixed-effects estimate of the main effect of interval (1).}
  \item{\code{I2}}{Fixed-effects estimate of the main effect of interval (2).}
  \item{\code{I3}}{Fixed-effects estimate of the main effect of interval (3).}
  \item{\code{R:I1}}{Fixed-effects estimate of the interaction (1).}
  \item{\code{R:I2}}{Fixed-effects estimate of the interaction (2).}
  \item{\code{R:I3}}{Fixed-effects estimate of the interaction (3).}
  \item{dev1}{Deviance for the model including the effect(s) of interest.}
  \item{dev2}{Deviance for the model excluding the effect(s) of interest.}
  \item{chisq_RI}{Chi-square value for the likelihood ratio test.}
  \item{p_RI}{Associated p-value.}
  \item{thresh.1|2}{First cut-point (threshold).}
  \item{thresh.2|3}{Second cut-point.}
  \item{thresh.3|4}{Third cut-point.}
  \item{thresh.4|5}{Fourth cut-point.}
  \item{thresh.5|6}{Fifth cut-point.}
  \item{thresh.6|7}{Sixth cut-point.}
}
}
\description{
Fit generalized additive mixed model to ratings data
}
\details{
Fits a generalized additive mixed model to the data and
  tests the specified effect (interaction or main effect) using a
  likelihood-ratio test using \code{mgcv::bam()}. If the
  interaction is to be tested, the following two models are
  compared.

\code{trating ~ R * (I1 + I2 + I3) +
   s(subj_id, bs = "re") +
   s(subj_id, R, I1, bs = "re") +
   s(subj_id, R, I2, bs = "re") +
   s(subj_id, R, I3, bs = "re") +
   s(stim_id, bs = "re") +
   s(stim_id, R, I1, bs = "re") +
   s(stim_id, R, I2, bs = "re") +
   s(stim_id, R, I3, bs = "re")}

\code{trating ~ R + I1 + I2 + I3 +
   s(subj_id, bs = "re") +
   s(subj_id, R, I1, bs = "re") +
   s(subj_id, R, I2, bs = "re") +
   s(subj_id, R, I3, bs = "re") +
   s(stim_id, bs = "re") +
   s(stim_id, R, I1, bs = "re") +
   s(stim_id, R, I2, bs = "re") +
   s(stim_id, R, I3, bs = "re")}

If the main effect is to be tested, then the following two models
are compared.

\code{trating ~ R * (I1 + I2 + I3) +
   s(subj_id, bs = "re") +
   s(subj_id, R, bs = "re") +
   s(stim_id, bs = "re") +
   s(stim_id, R, bs = "re")}

\code{trating ~ I1 + I2 + I3 + R:I1 + R:I2 + R:I3 +
   s(subj_id, bs = "re") +
   s(subj_id, R, bs = "re") +
   s(stim_id, bs = "re") +
   s(stim_id, R, bs = "re")}
}
\examples{
set.seed(62)
dat <- gen_data(40) 
fit_gamm(dat, TRUE) # test main effect

}
\seealso{
\code{\link{gen_data}}
}
