% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{flag_fake}
\alias{flag_fake}
\title{Flag the subdirectory as having simulated data}
\usage{
flag_fake(path)
}
\arguments{
\item{path}{Path to subdirectory.}
}
\description{
Flag the subdirectory as having simulated data
}
