% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{import_cjudgments}
\alias{import_cjudgments}
\title{Import Category Judgments From Multiple Files}
\usage{
import_cjudgments(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with the category judgment data.
}
\description{
Import Category Judgments From Multiple Files
}
