% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{import_cjudgments_simulated}
\alias{import_cjudgments_simulated}
\title{Import Category Judgments From Simulated Data}
\usage{
import_cjudgments_simulated(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with the category judgment data.
}
\description{
Import Category Judgments From Simulated Data
}
