% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{import_iratings}
\alias{import_iratings}
\title{Import Interest Ratings From Multiple Files}
\usage{
import_iratings(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with the interest ratings data.
}
\description{
Import Interest Ratings From Multiple Files
}
