% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{import_phase_info}
\alias{import_phase_info}
\title{Import Phase Information From Multiple Files}
\usage{
import_phase_info(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with phase information.
}
\description{
Import Phase Information From Multiple Files
}
