% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{import_phase_info_simulated}
\alias{import_phase_info_simulated}
\title{Import Phase Information From Simulated Data}
\usage{
import_phase_info_simulated(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with phase information.
}
\description{
Import Phase Information From Simulated Data
}
