% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{import_sessions}
\alias{import_sessions}
\title{Import Session Information From Multiple Files}
\usage{
import_sessions(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with session information.
}
\description{
Import Session Information From Multiple Files
}
