% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{import_sessions_simulated}
\alias{import_sessions_simulated}
\title{Import Session Information From Simulated Data}
\usage{
import_sessions_simulated(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with session information.
}
\description{
Import Session Information From Simulated Data
}
