% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{import_tratings}
\alias{import_tratings}
\title{Import Truth Ratings From Multiple Files}
\usage{
import_tratings(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with the truth rating data.
}
\description{
Import Truth Ratings From Multiple Files
}
