% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{import_tratings_simulated}
\alias{import_tratings_simulated}
\title{Import Truth Ratings From Simulated Data}
\usage{
import_tratings_simulated(path)
}
\arguments{
\item{path}{Directory containing response files.}
}
\value{
A table with the truth rating data.
}
\description{
Import Truth Ratings From Simulated Data
}
