% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{locate_data_files}
\alias{locate_data_files}
\title{Locate Raw Data Files from Longitudinal Illusory Truth Study}
\usage{
locate_data_files(path, full.names = TRUE)
}
\arguments{
\item{path}{Path to data files.}

\item{full.names}{If ‘TRUE’, the directory path is prepended to the
file names to give a relative file path.  If ‘FALSE’, the file
names (rather than paths) are returned.}
}
\value{
A character vector with the paths to the files.
}
\description{
Look in a subdirectory and find files containing the raw data.
}
\details{
Looks for files matching the regular expression
  \code{^[Pp][1-4]\\.[Cc][Ss][Vv]$} and performs basic
  error-checking.
}
