% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NE_items}
\alias{NE_items}
\title{Item Information from Nadarevic & Erdfelder 2014}
\format{
A data frame with 176 rows and 4 variables:
\describe{
  \item{item_id}{Unique stimulus (statement) identifier.}
  \item{statement}{Statement (in German).}
  \item{set}{Which set the statement belong to, used for counterbalancing.}
  \item{status}{Actual truth of the statement.}
}
}
\usage{
NE_items
}
\description{
A dataset describing the statements used as stimuli in Experiment 1 of Nadarevic & Erdfelder (2014).
}
\keyword{datasets}
