% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{normalize_path}
\alias{normalize_path}
\title{Get rid of trailing slash}
\usage{
normalize_path(path)
}
\arguments{
\item{path}{Directory name.}
}
\value{
Directory \code{name} without a trailing slash.
}
\description{
Get rid of trailing slash
}
