% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_sim}
\alias{power_sim}
\title{Run power simulations}
\usage{
power_sim(model, version, effect, nsubj, raw_eff, nruns,
  outfile = sprintf("\%s_\%s_\%s_\%04d_128_\%05d_\%0.2f_\%s_\%d.rds",
  model, version, effect, nsubj, nruns, raw_eff, Sys.info()[["nodename"]],
  Sys.getpid()))
}
\arguments{
\item{model}{Which type of model to fit: use 'lmem' for linear
mixed-effects model, 'clmm' for cumulative logit mixed-effects
model.}

\item{version}{Which assumption to make about the form of the
repetition-by-interval interaction; 'asymptote' for the
asymptotic version, 'single' for the effect-at-a-single-phase
version (see \code{\link{gen_data}}).}

\item{effect}{Which effect to test, the main effect (\code{'main'})
or the interaction effect (\code{'interaction'}).}

\item{nsubj}{Number of subjects.}

\item{raw_eff}{Size of the illusory-truth effect, in raw logit
units.}

\item{nruns}{How many simulations to run.}

\item{outfile}{Name of output file; \code{NULL} to return the
simulation results.}
}
\value{
Either the name of the outfile (if \code{outfile} is
  non-null) or the results of the simulation (a matrix containing
  results of \code{\link{fit_lmem}} or \code{\link{fit_clmm}}).
}
\description{
Run power simulations
}
\examples{
set.seed(62)
power_sim("lmem", "asymptote", "main", 40, .14, 1, NULL)
\dontrun{
power_sim("clmm", "asymptote", "main", 24, .14, 1, NULL) # takes ~10 minutes
}
}
