% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presentation_lists}
\alias{presentation_lists}
\title{Lists for counterbalanced stimulus presentation.}
\format{A data frame (tibble) with 1,536 rows and 6 variables:
\describe{

  \item{phase_id}{Phase (1-4) in which the stimulus will be presented.}

  \item{list_id}{Unique identifier of stimulus presentation list.}

  \item{task}{Which task the stimulus is presented in (rate
  interest or truth).}

  \item{stim_id}{Unique identifier of the stimulus.}

  \item{task_id}{Unique identifier of the task/stimulus combination.}

  \item{order}{Presentation order within the phase.}
}}
\usage{
presentation_lists
}
\description{
Lists for counterbalanced stimulus presentation.
}
\keyword{datasets}
