% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_iratings}
\alias{read_iratings}
\title{Read Interest Ratings From Response File}
\usage{
read_iratings(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A long table with the interest ratings.
}
\description{
Reads in the interest ratings from a single response file.
}
