% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_sessions}
\alias{read_sessions}
\title{Read Session Information From Response File}
\usage{
read_sessions(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A table with session information.
}
\description{
Read in session information from a single response file.
}
