% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_sessions_simulated}
\alias{read_sessions_simulated}
\title{Read Session Information From Simulated Response File}
\usage{
read_sessions_simulated(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A table with session information.
}
\description{
Read in session information from a single response file.
}
