% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{read_tratings}
\alias{read_tratings}
\title{Read Truth Ratings From Response File}
\usage{
read_tratings(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A long table with the truth ratings.
}
\description{
Reads in the truth ratings from a single response file.
}
