% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_tratings_simulated}
\alias{read_tratings_simulated}
\title{Read Truth Ratings From Simulated Data}
\usage{
read_tratings_simulated(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A long table with the truth ratings.
}
\description{
Reads in the truth ratings from a single response file.
}
