% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_analysis_sim}
\alias{reproduce_analysis_sim}
\title{Simulate the Analysis for Longitudinal Illusory Truth Study}
\usage{
reproduce_analysis_sim(
  path,
  outfile = "analysis.html",
  recipe = FALSE,
  infile = NULL
)
}
\arguments{
\item{path}{Path to a subdirectory containing the preprocessed
anonymized (simulated) data files (see
\code{\link{simulate_resp_files}}).}

\item{outfile}{Path to the HTML output file.}

\item{recipe}{Include instructions on how to reproduce the analysis.}

\item{infile}{Path to the R Markdown script; \code{NULL} to use the
built-in script.}
}
\description{
Runs the main analysis for
\insertCite{Henderson_Simons_Barr_2021;textual}{truthiness} on
simulated data.
}
\details{
Runs R Markdown script on the data in the provided
  subdirectory and renders the HTML report to \code{outfile}. The
  master R Markdown script can be accessed using:

\code{rmarkdown::draft("analysis.Rmd",
                       "illusory-truth-analysis-sim", "truthiness")}

Note that this script can take *very* long to run, depending on the
size of the simulated dataset, the number of processing cores, and
the computational power of the hardware.
}
\examples{
td_raw <- tempfile()  # temp dir for raw data
td_anon <- tempfile() # temp dir for preprocessed data

## simulate data and preprocess it

set.seed(62)
simulate_resp_files(40, path = td_raw, overwrite = TRUE)
report <- preprocess_simulated(td_raw, td_anon)

\dontrun{
reproduce_analysis_sim(td_anon)
}

## clean up
file.remove(report)
unlink(td_raw, TRUE, TRUE)
unlink(td_anon, TRUE, TRUE)

}
