% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sessions}
\alias{sessions}
\title{Session Information from Truth Trajectory Study}
\format{
A data frame (tibble) with 631 observations on 17 variables:
\describe{
  \item{ID}{Participant identifier.}
  \item{list_id}{Stimulus list identifier.}
  \item{Age}{Age of participant in years.}
  \item{Gender}{Gender of participant.}
  \item{Nationality}{Nationality of participant.}
  \item{NativeLang}{Native language of participant.}
  \item{keep}{Logical value, whether to keep (TRUE) or exclude (FALSE) participant.}
  \item{excl_phase}{Phase in which exclusion occurred (or NA).}
  \item{excl_reason}{Reason for exclusion (or NA).}
  \item{chk_noduplicates}{Whether there were no duplicated sessions.}
  \item{chk_consent_all}{Whether participant gave consent for all phases.}
  \item{chk_native}{Whether participant is a native speaker of English.}
  \item{chk_nocheat}{Whether participant never looked up answers.}
  \item{chk_dur_all}{Whether all phase durations were within an acceptable range.}
  \item{chk_noflatline}{Whether the participant did not produce 'flatline' responses.}
  \item{chk_anydata}{Whether there is any ratings data for this participant.}
  \item{chk_notmanex}{Whether the participant is not manually excluded.}
}
}
\usage{
sessions
}
\description{
A dataset representing information about each participant who took part in the study.
}
\details{
The logical variable keep is a boolean AND of the
  variables `chk_noduplicates`, `chk_consent_all`, `chk_native`,
  `chk_nocheat`, `chk_dur_all`, `chk_noflatline`, `chk_anydata`,
  and `chk_notmanex`.
}
\keyword{datasets}
