% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{simulate_category_guess}
\alias{simulate_category_guess}
\title{Simulate guessing during the categorization task.}
\usage{
simulate_category_guess(nruns = 10000)
}
\arguments{
\item{nruns}{Number of simulation runs.}
}
\value{
A vector of length \code{nruns} with the number of correct guesses.
}
\description{
Run simulations tabulating the number correct under guessing in the categorization task.
}
