% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stimulus_categories}
\alias{stimulus_categories}
\title{Target categories for stimulus items}
\format{
A data frame with 170 observations on 3 variable:
\describe{
  \item{stim_id}{Unique identifier of the stimulus.}
  \item{choice}{Category number, allows for more than one correct categorization.}
  \item{category}{Name of category.}
}
}
\usage{
stimulus_categories
}
\description{
A dataset representing the category or categories a stimulus belongs to for the categorization task.
}
\keyword{datasets}
