% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stimulus_conditions}
\alias{stimulus_conditions}
\title{Lookup table identifying stimulus conditions across lists.}
\format{A data frame (tibble) with 1,024 rows and 4 variables:
\describe{

  \item{list_id}{Unique identifier of stimulus presentation list.}

  \item{stim_id}{Unique identifier of the stimulus.}

  \item{repetition}{Whether the statement was repeated or novel.}

  \item{interval}{Interval (phase) in which the truth-rating was
  performed (immediate = 1, 1 day = 2, 1 week = 3, 1 month = 4).}

}}
\usage{
stimulus_conditions
}
\description{
Lookup table identifying stimulus conditions across lists.
}
\keyword{datasets}
