% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthiness.R
\docType{package}
\name{truthiness}
\alias{truthiness}
\title{truthiness: Longitudinal Study of the Illusory Truth effect}
\description{
The truthiness package provides functions and data associated with
a longitudinal study of the Illusory Truth effect conducted by
\insertCite{Henderson_Simons_Barr_2021;textual}{truthiness}.
}
\references{
\insertAllCited{}
}
