% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{validate_filenames}
\alias{validate_filenames}
\title{Validate Filenames in Directory}
\usage{
validate_filenames(path)
}
\arguments{
\item{path}{Path to the files.}
}
\description{
Make sure all the files needed for the analysis are present in a
given directory.
}
\details{
Output files from the study must match the pattern
  \code{PXLY.csv} where X is phase number (1-4) and Y is list
  number (1-8).
}
