% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{warn}
\alias{warn}
\title{Warn About Simulated Data}
\usage{
warn(subdir)
}
\arguments{
\item{subdir}{Subdirectory with the anonymized data.}
}
\description{
Check whether the data in \code{subdir} is simulated data and
generate a warning to include in an R Markdown document.
}
