% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{derive_fixed}
\alias{derive_fixed}
\title{Derive Fixed-Effects Parameters from Phase-by-Phase Effects}
\usage{
derive_fixed(phase_eff)
}
\arguments{
\item{phase_eff}{A four-element vector specifying the illusory
truth effect at each of the four testing phases (in log odds
units).}
}
\value{
Vector with eight elements containing the fixed effects
  coefficients for deviation-coded predictors.
}
\description{
Derive Fixed-Effects Parameters from Phase-by-Phase Effects
}
