% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_files.R
\name{normalize_path}
\alias{normalize_path}
\title{Get Rid of Trailing Slash}
\usage{
normalize_path(path)
}
\arguments{
\item{path}{Directory name.}
}
\value{
Directory name without a trailing slash.
}
\description{
Remove extra trailing slash from file path.
}
