% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_equiv}
\alias{power_equiv}
\title{Power Simulation For Equivalence Test}
\usage{
power_equiv(
  phase_eff,
  delta,
  target_effect,
  nsubj,
  nruns,
  outfile = sprintf("\%s_\%s_\%s_\%04d_128_\%05d_\%s_\%d.rds", ifelse(is.null(delta),
    "model", "equiv"), if (is.null(delta)) {    
    sprintf("\%0.2f~\%0.2f~\%0.2f~\%0.2f~x.xx", phase_eff[1], phase_eff[2], phase_eff[3],
    phase_eff[4]) } else {     sprintf("\%0.2f~\%0.2f~\%0.2f~\%0.2f~\%0.2f",
    phase_eff[1], phase_eff[2], phase_eff[3], phase_eff[4], delta) }, target_effect,
    nsubj, nruns, Sys.info()[["nodename"]], Sys.getpid())
)
}
\arguments{
\item{phase_eff}{A four-element vector, each element of which
specifies the illusory truth effect at the corresponding phase,
on the log odds scale (see \code{\link{gen_data}}).}

\item{delta}{Smallest (raw) effect size of interest, on log odds
scale; \code{NULL} to store fitted model object.}

\item{target_effect}{Which effect to test, the main effect
(\code{'main'}) or the interaction effect (\code{'interaction'}).}

\item{nsubj}{Number of subjects.}

\item{nruns}{How many simulations to run.}

\item{outfile}{Name of output file; \code{NULL} to return the
simulation results.}
}
\value{
Either the name of the outfile (if \code{outfile} is
  non-null) or the results of the simulation (a matrix containing
  results of \code{\link{fit_lmem}} or \code{\link{fit_clmm}}).
}
\description{
Power Simulation For Equivalence Test
}
\examples{
set.seed(62)
power_equiv(c(0, .14, .14, .14), "main", 40, 1, NULL)
\dontrun{
power_equiv(c(0, .14, .14, .14), "main", 24, 1, NULL) # takes ~10 minutes
}
}
