% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process and Anonymize Response Data}
\usage{
preprocess(
  path,
  outpath = paste0("anon-", Sys.Date()),
  report = paste0(basename(outpath), "-preprocessing.html")
)
}
\arguments{
\item{path}{Path to the directory containing raw response files.}

\item{outpath}{Path to the directory where anonymized data will be saved.}

\item{report}{Filename of the HTML preprocessing report.}
}
\value{
Path where the files were written (\code{outpath}).
}
\description{
Pre-process and Anonymize Response Data
}
\details{
Loads in the data from the raw response files and writes
  out non-anonymized, pre-processed versions to the current working
  directory, as well as anonymized versions to the directory
  specified by \code{outpath}. The structure of the files is
  described by the function \code{\link{codebook}}.
}
