% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process and Anonymize Response Data}
\usage{
preprocess(inpath, outpath = paste0("anon-", Sys.Date()), overwrite = FALSE)
}
\arguments{
\item{inpath}{Path to the directory containing raw response files.}

\item{outpath}{Path to the directory where anonymized data will be saved.}

\item{overwrite}{Whether to overwrite the anonymized data.}
}
\value{
Path where the files were written (\code{outpath}).
}
\description{
Pre-process and Anonymize Response Data
}
\details{
Loads in the data from the raw response files and writes
  out non-anonymized, pre-processed versions to the current working
  directory, as well as anonymized versions to the directory
  specified by \code{outpath}. The anonymized output files have the
  following structure.

\code{ANON_sessions.csv}: A table with one row for each participant
and four columns:

\describe{
  \item{ID}{Unique participant identifier.}
  \item{list_id}{Identifier of stimulus list.}
  \item{chk_native}{Whether the participant was a native English speaker.}
  \item{chk_nocheat}{Whether the participant did not look up answers.}
}

\code{ANON_phases.csv}: A table with one row for each participant
and five columns:

\describe{
  \item{ID}{Unique participant identifier.}
  \item{phase_id}{Identifier of phase (1-4).}
  \item{duration_secs}{Duration of phase in seconds.}
  \item{chk_finished}{Whether the participant completed the phase.}
  \item{chk_dur}{Whether the phase was completed within a reasonable duration.}
}

\code{ANON_interest.csv}: A table with one row for each interest
rating and four columns:

\describe{
  \item{ID}{Unique participant identifier.}
  \item{stim_id}{Unique stimulus identifier.}
  \item{order}{Order of presentation.}
  \item{irating}{Interest rating (0-10).}
}

\code{ANON_ratings.csv}: A table with one row for each truth rating
and five columns.

\describe{
  \item{ID}{Unique participant indentifier.}
  \item{phase_id}{Identifier of phase (1-4).}
  \item{stim_id}{Unique stimulus identifier.}
  \item{order}{Order of presentation.}
  \item{trating}{Truth rating (1-7).}
}
}
