% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_cjudgments}
\alias{read_cjudgments}
\title{Read Category Judgments From Raw Data File}
\usage{
read_cjudgments(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A long table with the interest ratings.
}
\description{
Reads in the category judgments from a single response file
containing phase data.
}
