% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_analysis}
\alias{reproduce_analysis}
\title{Reproduce the Master Data Analysis}
\usage{
reproduce_analysis(
  path,
  outfile = "my_analysis.html",
  inferential = FALSE,
  simulated = FALSE,
  instructions = FALSE
)
}
\arguments{
\item{path}{Path to the anonymized data files.}

\item{outfile}{Path to the HTML output file.}

\item{inferential}{Whether to run inferential statistics.}

\item{simulated}{Whether the analysis is performed on simulated
data (TRUE) or real data (FALSE). In the case of simulated data,
the compiled document will contain warnings that the data are not
real.}

\item{instructions}{Include instructions on how to reproduce the analysis.}
}
\description{
Reproduce the Master Data Analysis
}
\details{
Runs the R Markdown script on the data in the provided
  subdirectory and renders the HTML report to \code{outfile}. The
  master R Markdown script can be accessed using:

\code{rmarkdown::draft("master_script.Rmd",
                       "illusory-truth-analysis", "truthiness")}
}
