% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_analysis}
\alias{reproduce_analysis}
\title{Reproduce the Analysis for Longitudinal Illusory Truth Study}
\usage{
reproduce_analysis(
  outfile = "analysis.html",
  refit = FALSE,
  savefig = FALSE,
  recipe = FALSE,
  infile = NULL
)
}
\arguments{
\item{outfile}{Path to the HTML output file.}

\item{refit}{Whether to re-fit the cumulative link mixed model
\code{TRUE} or to use the built-in model fits (\code{FALSE}). Due
to the extremely time-consuming nature of model estimation, the
default is set to \code{FALSE}.}

\item{savefig}{Whether to save the two plots as separate PNG files
(\code{means_plot.png} and \code{validation_plot.png}).}

\item{recipe}{Include instructions on how to reproduce the analysis.}

\item{infile}{Path to the R Markdown script; \code{NULL} to use the
built-in script.}
}
\value{
Path to the rendered HTML report.
}
\description{
Re-run the analysis for the
\insertCite{Henderson_Simons_Barr_2021;textual}{truthiness}
longitudinal truth study.
}
\details{
Runs R Markdown script containing the analysis code. The
  analysis is performed on the built-in preprocessed anonymized
  data (documented in \code{\link{truth_trajectory_data}}). The
  script output is rendered as an HTML report, specified by
  \code{outfile}. Although it is not necessary to do so, the master
  R Markdown script for processing real data can be accessed using

\code{rmarkdown::draft("analysis.Rmd",
                       "illusory-truth-analysis", "truthiness")}
}
\examples{
tf <- tempfile(fileext = ".html")
result <- reproduce_analysis(tf)
\dontrun{browseURL(result)}

\dontrun{
## to re-fit the data (takes VERY long):
result <- reproduce_analysis(refit = TRUE)
}

}
\seealso{
\code{\link{reproduce_analysis_simulated}}
}
