% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce_analysis}
\alias{reproduce_analysis}
\title{Reproduce the Master Data Analysis}
\usage{
reproduce_analysis(path, outfile = "analysis.html", recipe = FALSE,
  infile = NULL)
}
\arguments{
\item{path}{Path to the anonymized data files.}

\item{outfile}{Path to the HTML output file.}

\item{recipe}{Include instructions on how to reproduce the analysis.}

\item{infile}{Path to the R Markdown script; \code{NULL} to use the
built-in script.}
}
\description{
Reproduce the Master Data Analysis
}
\details{
Runs R Markdown script on the data in the provided
  subdirectory and renders the HTML report to \code{outfile}. The
  master R Markdown script can be accessed using:

\code{rmarkdown::draft("master_script.Rmd",
                       "illusory-truth-analysis", "truthiness")}
}
