% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{run_equiv}
\alias{run_equiv}
\title{Run equivalence test}
\usage{
run_equiv(.data, main_effect = FALSE, delta = 0.14)
}
\arguments{
\item{.data}{Data frame.}

\item{main_effect}{Whether to run the test for the main effect
(TRUE) or the interaction (FALSE).}

\item{delta}{Smallest (raw) effect size of interest (log odds scale).}
}
\description{
Run equivalence test
}
